tdf#144200 Add CFBundleSupportedPlatforms key with MacOSX as its only platform
The necessity of including the CFBundleSupportedPlatforms key in an
application that does not include the Intel architecture is described
in the following post:

https://github.com/microsoft/vscode/issues/121169#issuecomment-820653677

The above post is quoted below:

"Its just an issue with which attribute the storage info queries to
decide the kind, the output of system_profiler SPApplicationsDataType -xml
shows the arch_kind is arch_ios. I think these look for
CFBundleSupportedPlatforms attribute and if not present default to the ios
kind for arm64 executables.

Given the universal and intel labels show up correctly, and considering
the output of other apple silicon apps in the image posted. I don't think
this needs a fix, but leaving it open for anyone to address as a good
first issue."

Change-Id: Ic96ce4407e060fbb0cb8c7b0b72dcf8f61dade19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184738
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
(cherry picked from commit 562a40f453caa003187ff28c7798647d9e0b4ded)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184769
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>