tdf#166319 sw interdependent redlines: handle format on top of insert
Open the bugdoc, the DOCX file has an insert with a format on top of the
middle, but the format is lost in Writer.

The first problem is that we tried to create an insert on top of format
(while in DOCX the format is always inside the insert/delete), fix that
by changing the order of redline creation in
DomainMapper_Impl::CheckRedline(). The original motivation is no longer
relevant after commit b3567d7e1f08df7c050d3cec86cbb200af558172
(tdf#165322 sw: apply formatting for multiple redlines in SwRedlineItr,
2025-02-25) Writer core to handle the rendering of format on top of
insert/delete.

The other problem is that in case we try to append a format redline on
top of insert ("AB" is an insert and the format would affect "B"), then
we simply deleted the (candidate) format redline -- change this to
handle this similar to delete on top of insert.

CppunitTest_sw_ooxmlexport8's testN830205 would now create an unsorted
redline table on file open, so make sure SwDoc::TextToTable() resorts
the table after splitting the redlines.

Change-Id: Iada13204892ffccfe09bc2c28bf0c563db6e1ab9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184752
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins