vcl: more fine grained handling of error in PNG reader
Handle if the error happens at init, reading and after reading
differently. At init we don't need to handle the bitmap as nothing
has yet been writtent to it. Also after reading we can handle it
like no error happened, because we have read the data completely
but the error happened after that (like missing IEND chunk for
example).

Change the error and warning messgaes when reading from the stream.

Change-Id: I927b8679f168da43e8e6e00663ff551064b870c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184811
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>