CEOO: CellEditOnOverlay
Instead of using EditView's Paint directly to Window I
changed this now to using the Overlay that is available in
all applications. This allows to have less repaints of the
Calc View since the Overlay has a copy of the area it is
working overlayed.
It also allows to get away from one of the last areas where
XOR selection was used since EditView/EditEngine have no
better way to do this. Selection is now visualized using the
OS'es choosen SelectionColor, as in all other apps.
I could also get rid of the flush() of the OverlayManager
that was needed before and caused some problems (see
tdf#165621 and associated tasks).
A failing UnitTest showed that OverlayManagerBuffered needs
to be aware when no CompleteRedraw was done yet and thus no
background data exists and refresh makes no sense. This may
happen when there is no redraw/UI e.g. UnitTests, but still
an OverlayManager is used and together with a Reschedule
(many possible reasons) that refresh might be triggered.
Made access to EditView in OverlayObject no longer dependent
on ActiveView, that is not always the one with the EditView.
Now using a combination of ScTabView and ScSplitPos for
accesses. Asserting now if no EditView -> is mandatory.
Continuing going through CppunitTest_sc_tiledrendering UTs,
one more identified.
Have to re-activate EditViewInvalidate rect forwarding
since the Invalidate is needed for sc tiled rendering
tests, but not for sd tiled rendering tests. Added that
to the impl of EditViewCallbacks using
LibreOfficeKit::isActive().

Change-Id: I41f5448bfbabfaae4858c7617478771053c2cd77
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184830
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>