sw redline: extract a PreAppendInsertRedline() from AppendRedline()
sw::DocumentRedlineManager::AppendRedline() is 1158 lines long, which
makes it very hard to reason about the body of the function.

3 extractable pieces of it are the blocks which tweak the to-be-inserted
redline and possibly existing redlines of the redline table based on the
type of the to-be-inserted redline (insert, delete, format).

Extract the insert redline part into a new function, which alone is 261
lines long.

To keep this mechanical, pass all variables by non-const ref, then make 2
selected variables const, since those are only read.

Change-Id: I105524e010bab92afdf54bbba5b33dccdc5900dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184980
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>