basicide a11y: Move basctl::createTextWindowPeer logic to caller
Drop the helper method basctl::createTextWindowPeer
that calls the TextWindowPeer ctor, and instead
call the ctor directly from
EditorWindow::GetComponentInterface.

In order for the class to be known there, move it
to the anonymous namespace in the same source file.

This is an intermediate step towards
getting rid of TextWindowPeer altogether.

Change-Id: I58111af8fadb45c6471af942b300665f3df189c9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184837
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>