tdf#130857 qt weld: Evaluate "label" property for tool buttons
Evaluate the GtkToolButton::label property for "GtkToolButton"
objects in .ui files, and set the text for the QToolButton based
on it.

This e.g. makes the Orca screen reader announce the "Paste" and
"Special Characters" tool buttons in the spelling dialog as such
instead of just saying "button" when they receive focus in a
WIP branch adding support for that dialog with the qt6 VCL plugin
and SAL_VCL_QT_USE_WELDED_WIDGETS=1.

Change-Id: I35a1c6940881d47f6d788144ef221b089cbd67bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184847
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>