tdf#130857 qt weld: Assert widgets exist, drop fallback
In QtInstanceBuilder::weld_*, add asserts that the
widgets for the passed ID exist and drop the fallback
of returning an empty std::unique_ptr if they don't.

Dialog implementations generally don't handle null return
values anyway, and if these methods get called with an
ID for which there isn't a corresponding widget, this needs
to be investigated, so rather fail here early than
due to a null dereference later.

Change-Id: I6d018a4fe8b9ad19a01b7072281d5eb215ab8ef0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184875
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins