tdf#130857 qt weld a11y: Support custom DrawingArea a11y
Add support for using a custom accessibility
interface/implementation for weld::DrawingLayer
using native Qt widgets with
SAL_VCL_QT_USE_WELDED_WIDGETS=1:

Handle the a11yref/uno::Reference<XAccessbible> param
passed to QtInstanceBuilder::weld_drawing_area by
implementing logic to set a custom QAccessibleInterface
implementation for the native QLabel that is used
in the QtInstanceDrawingArea:

Delete the QtAccessibleInterface currently set for
the label and create a QtAccessibleWidget that basically
wraps the XAccessible and is already used for non-weld
objects, too. QtAccessibleWidget::customFactory is already
registered as an a11y factory for Qt, so extend it to
evaluate a QObject property to transport the custom
accessible interface implmentation into that method
from QtAccessibleWidget::setCustomAccessible.

This, together with some further changes in a WIP
branch implementing support for the spelling dialog
result in the "Not in dictionary" edit view becoming
accessible, i.e. expose its paragraph children on
the a11y layer, support the AT-SPI Text interface
and emit caret-changed events when moving the cursor
within the edit view.

Change-Id: I4686f1d6c5862fea38af626facff785ed5e9f9dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184879
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>