tdf#130857 qt weld a11y: Drop 2 DrawingArea a11y asserts
No longer assert in QtInstanceDrawingArea::get_accessible_parent
and QtInstanceDrawingArea::get_accessible_relation_set, but just
let them return nullptr.

With

    Change-Id: I4686f1d6c5862fea38af626facff785ed5e9f9dd
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu May 1 23:21:55 2025 +0200

        tdf#130857 qt weld a11y: Support custom DrawingArea a11y

in place, those now get called for the
SAL_VCL_QT_USE_WELDED_WIDGETS=1 case in
a WIP branch adding support for the spelling
dialog when the Orca screen reader is running,
because QtAccessibleWidget::parent and
QtAccessibleWidget::relations call them.

However, QtAccessibleWidget::parent already
falls back to the native QWidget/QObject
hierarchy to find the accessible parent, which
works for this case, too. (Adjust the comment
a bit to mention the new scenario of native weld::Widget
instances as well.)

Relations should generally also be handled directly
by QtAccessibleWidget, not the underlying XAccessible,
as they are set in the .ui file which is evaluated
by QtBuilder. If something is still missing there,
then the logic still needs to be refined elsewhere.
Come back to this later as needed.

gtk3's GtkInstanceDrawingArea implementations also
don't return anything particular here, but in fact
simply trigger an assert when called, mentioning
that this shouldn't be reached for the implementation
there as everything is handled natively on the ATK
layer.

Change-Id: I6c9fa487614b5cb984888842e6c137801ab97b8a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184880
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>