tdf#130857 qt weld a11y: Add custom a11y to registry
Insert the mapping between XAccessible and the associated
QLabel to QtAccessibleRegistry for the case where a custom
a11y object is used as implemented in previous commit

    Change-Id: I4686f1d6c5862fea38af626facff785ed5e9f9dd
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu May 1 23:21:55 2025 +0200

        tdf#130857 qt weld a11y: Support custom DrawingArea a11y

This makes sure that the existing QtAccessibleWidget
is used for the XAccessible later, instead of creating
another one/a QtXAccessible later.

See also

    commit 61c0c1286dbd9015809ba8ee5ee687b612438bef
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Aug 24 13:47:25 2022 +0200

        qt a11y: Remember associated QObject also for non-QtXAcccessible case

for a similar change just a few lines further down, which
also has more details.

This commit makes the paragraph child in the spelling dialog
report the edit view/drawing area accessible as its parent on the
a11y layer as it should, and thus also makes Accerciser
highlight the correct area on screen when clicking on them,
while that was not the case previously (at least on Wayland)
with SAL_VCL_QT_USE_WELDED_WIDGETS=1 in a WIP branch adding
support for using native Qt widgets for the spelling dialog.

Change-Id: I57a8494152b867a999aecff0e94c2e7b4373881e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184882
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>