Revert "tdf#130857 qt weld: Assert widgets exist, drop fallback"
It turns out that there are use cases where IDs are
used for which no widget exists in the .ui file.

At leasts the SmCategoryDesc ctor doesn't use hard-coded
widget IDs but generates them dynamically and includes
`nCategoryIdx`, so the resulting ID may or may not
have an associated widget, and the null case is handled there.

This reverts commit 960703ce7fe55af062bd33e6bc9f6f33bfd193e8.

Change-Id: I760f1ac82c5e4e22409fe0ab99309e6b23c37aa7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184895
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>