tdf#165858 pdf: use the correct widget to dissble the forms options
"formsframe" widget was a grid that previously contained the forms
options only, but after restructuring the "forms_frame" is actually
a frame that contains all the forms option and checkbox to enable
writing of the forms. A different "forms_options_grid" is used to
contain only the form options.

This was changed in the .ui file, but it was not changed correctly
in the code (only "formsframe" was renamed to "forms_frame").

The change fixes the code to use "forms_options_grid" instead of
"forms_frame" and renames the variables so now the correct widget
is used.

Change-Id: I15f967575d1a53ab9545e68867c48abefbc6b0dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184910
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 86e8abc2b171f56511787eda28f1604a0ee6d000)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184932
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>