ScriptForge (UI) extend scope of CreateBaseDocument
The
    ui.CreateBaseDocument()
method accepted so far to create Base documents
either embedding HSQLDB or FIREBIRD databases
or linked to a Cals spreadsheet.

It is now extended to make the new Base document
finding its datasource in an external Firebird
database file.

The previous signature
    ui.CreateBaseDocument(FileName, EmbeddedDatabase,
        CalcFileName)
is replaced by
    ui.CreateBaseDocument(FileName, EmbeddedDatabase,
        DataFileName, CalcFileName)
=> The CalcFileName argument is deprecated. When present,
   it is copied to DataFileName.
The EmbeddedDatabase argument may take next values:
    HSQLDB (default), FIREBIRD, FIREBIRD_EXTERNAL or CALC.

The change is valid for both Basic and Python scripts.
The documentation should be upgraded.

Change-Id: I5daf418f50739baca6be899eb1444d1b99cf89aa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184935
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins