tdf#166319 sw interdependent redlines: handle accept of insert under format
Open the bugdoc, move the cursor inside "AAA", accept the change, the
result is that "BBB" still has a change, which is marked as a "format on
top of insert", while the intent was to accept the insert.

A similar use-case is the same with "delete on top of insert", and that
one works since commit 52fa7aed48632166e064e6a227e034f0981c4205
(tdf#157662 SW: redline: accept/reject done for all parts, 2023-10-14).

Fix the problem by extending
sw::DocumentRedlineManager::AcceptRedlineRange(), so it considers
format-on-insert the same as delete-on-insert, so all 3 redlines get
their insert part accepted.

The result is that that the insert-only redlines are deleted and the
insert-and-format redline is downgraded to be a simple format redline.

Change-Id: I2fdcc906d40c7e94d5446e8912390fe814a5f84c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184983
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>