cool#11826 desktop lok: inform new view about already painted areas
Open the Calc bugdoc, sheet 1 is visible. Open a 2nd view, still on
sheet 1. Switch to sheet 2 in view 1, change the call, the sheet 1 in
view 2 doesn't update.

What happens is that view 1 performs its rendering of sheet 1, then view
2 is created, so once sheet 2 is changed in view 1, we think that there
is no need to invalidate view 2, since it was never painted, since
commit b95a6816555396661d90da02e844d2848b1da3bf (cool#11254 desktop lok:
avoid invalidations if no tiles are sent, 2025-03-05).

Fix the problem by extending doc_registerCallback() to take the info
about already painted areas on view creation: now that the view creation
is handled and the view update is done in
LibLODocument_Impl::updateViewsForPaintedTile(), the info about already
painted tiles is again synchronized in all views with the same view
render state.

Also mention CanonicalViewId in comments, because that's term we use for
views with the same view render state, so git grep can find both
locations.

Change-Id: I7698830e1070e0bbe0c186d41ce83c8e22eb7049
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184977
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>