cool#11785 sw lok: take visible area into account when running idle jobs
Load a Writer document that is large enough that loading takes more than
a second, then spellcheck will be done for the entire document in LOK
mode, and only then we'll render the first tile.

This was like this since commit 6a490e8834b276359b9ada2ebc95bc86a2dbe759
(cool#11785 sw lok: only start idling after the idle jobs timer stopped,
2025-04-29), previously we always did a full spellcheck after load,
before rendering a first tile.

Fix the problem by reverting the previous fix (that restored a 1 sec
window in time to rendere first tiles) and instead improve
SwLayIdle::DoIdleJob(), so in case it wants to only do a job for the
visible area, then it takes the LOK visible area into account, similar
to how idle layout does this in SwLayIdle::DoIdleJob().

This probably means m_aFireIdleJobsTimer can be fully removed in the
future, but that's not yet done here.

Change-Id: I676137daf78cb4fd3be90f474e60213846908fef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185039
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>