make libeot compile on Windows
Compile for Windows with LO infrastructure (not the only external lib
where we do this).

Remove unneeded include of <err.h> in libeot.c that isn't used, but
Windows doesn't have <err.h> - it's a BSD / Linux thing.

Only compile libeot.la and not the eot2ttf binary (which uses err.h so
it doesn't compile on Windows).

Change-Id: I26afb2b7bcf0bdf0d2ce0bdbcfceb10599d599e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185057
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185075
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>