CppunitTest_sw_layoutwriter3: fix en-GB vs en-US mismatch
The test code at testTdf126154_minimum_shrinking and testTdf126154
checked if en-US hyph data is available, but the documents worked with
an en-GB locale, so it could happen that you had en-US installed and
then the test failed if en-GB was not installed.

Adapt the test documents to the test code to fix the inconsistency.

Change-Id: I826db1bdb19bb383ee8e992eac64495670889d7d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185090
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>