gen: Drop complexity as string is always empty anyway
While this does interesting checks for the first arg,
`aDisplayName` will always be an empty string as
an empty/default-constructed Any is used when assigning a
value using `operator>>=`.

Therefore, drop the extra logic that doesn't
have any effect, and go with the default param of
an empty OUString for SelectionManager::get.

Change-Id: Ibc1b6637e3def37f96ab074bde7fe40b60e330c3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185104
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>