tdf#164380 qt: No longer use vcl::X11DnDHelper, XInitialization
Stop using vcl::X11DnDHelper to initialize
QtDragSource and QtDropTarget.
vcl::X11DnDHelper uses the XInitialization interface
to pass the display connection and window via
XInitialization::initialize.

However, the display connection isn't used, and
the window handle is already known, so instead
pass those directly to the QtDragSource and QtDropTarget
ctors, and stop supporting the XInitialization interface
altogether.

The parameters passed to XInitialization::initialize
are documented for the X11DragSource and X11DropTarget
services in their .idl files, but those are not relevant
here in the Qt implementation, because this one is only
supporting the XDragSource and XDropTarget
interfaces and not reporting the service names including
the "11" in the name (i.e. specific to the X/X11 window system)
in the name as supported.

Change-Id: Ia8a4d305dfcb3916ba44596b31614e2467701514
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185108
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins