tdf#164380 gtk: No longer use vcl::X11DnDHelper, XInitialization
This is the gtk3/gtk4 version of

    Change-Id: Ia8a4d305dfcb3916ba44596b31614e2467701514
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri May 9 14:27:58 2025 +0200

        tdf#164380 qt: No longer use vcl::X11DnDHelper, XInitialization

, see that commit's message for more details.

For the GTK based VCL plugins, leave the default
ctor for both classes in place, because that one is
also called from elsewhere; but there, XInitialization::initialize
is not called at all, so leave the logic for the default
ctor unchanged (and call it from the new ctor taking
one param).

No change in behavior intended, and drag'n'drop to/from
Firefox to Writer was still working as expected in a quick test.

Change-Id: Icade8e26f055a7efd75a1aa1a0d3e058ab09daaa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185109
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>