tdf#164380 vcl: Move X11DnDHelper to gen, drop return value
By now, vcl::X11DnDHelper is only used in the gen VCL
backend, so move the helper function there and rename
it to just "InitializeDnD".

Drop the return value, as its the same as an input
parameter (and thus the callers already have a reference
to the object of interest).

Resolve disambiguity for SelectionManagerHolder and DropTarget
(which inherit from XInterface via multiple paths in the class
hierarchy) in a way that already prepares for returning
a more specific interface from
SalInstance::ImplCreate{DragSource,DropTarget}
in the future.

Change-Id: I16f2be9876cbb6cc7d93b40820460dd1684d1e2c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185110
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>