tdf#163015 FILEOPEN: Opening specific XLSX file takes long time
we spend most of our time inserting into the output sorted_vector,
which ends up needing a lot of data movement when the output
vector becomes large.
Re-arrange the data structures so we can do this in one pass,
and instead of inserting into the output vector, we can just
append to it.

Change-Id: If729e8ca1f2e6b5cf546c17331f1f3716674f2cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185099
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins