qt: Use dedicated SystemEnvData::pSalFrame for frame
This is basically the Qt version of previous commit

    Change-Id: I569e147182776c2d78f2269b1dc5eb37d009e3bf
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat May 10 00:09:51 2025 +0200

        gtk: Use dedicated SystemEnvData::pSalFrame for frame

for the GTK based VCL plugins.

In addition, this extends QtFrame::FillSystemEnvData
to actually set SystemEnvData::pSalFrame (as unlike for
the gtk VCL plugins, that wasn't done earlier yet)
and pass the frame when calling that method from
the QtFrame ctor.

Change-Id: I17282ce785ce48ffd2e4346a3108d12ab51c2786
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185121
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins