tdf#166511 sw: unify export of DrawingML object IDs
Both docPr and cNvPr have IDs referring to DrawingML objects,
and are supposed to be unique within the document.

These IDs were handled in different parts of the code during export.
7460e4f4a7b15cc7984adf65bc17e3d580413224 added a workaround,
likely to keep them unique, but together with
4b0fa253a4540f5461397815d290586f9ddabe61 this caused a clash
sometimes, and the result couldn't be opened in Word afterwards.

Use the same counter for both IDs. It needs to start from 1, as 0
ID makes at least grouped shapes invalid.

Change-Id: I5e7b9855bc88a61841dfcd3d41599a4ce444eb40
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185191
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185311
Reviewed-by: Andras Timar <andras.timar@collabora.com>