tdf#166229 sw: make bullet style change again with list indentation level
Create a new document, turn on bullets, create a next paragraph,
increase the indent level, the same bullet character is used for the
inner level, while this used to be different.

This went wrong in commit 626357f53c934e7f57dc80c3c83ad080767961f3 (Use
configured bullet symbol when clicking bullets button, 2024-07-11),
which changes Writer to take the bullet characters from the common
DefaultBullets settings, because it wanted Writer bullets to respect
that setting. That's OK, but it always took the bullet character
configured for the first level, also the size of the characters were
inconsistent (first is significantly smaller than 2nd, 3rd, 4th).

Fix the problem by taking the current level index into account in
SwWrtShell::NumOrBulletOn(), unless the configured bullet character list
is shorter than the amount of available Writer indentation levels. Also
change the DefaultBullets item values to match what was the default in
Writer.

Likely this is not controversial, since it seems Impress has its
own defaults: Writer has 3 characters cycled through 10 levels, Impress
seems to be switching between "•" and "-" for the first 4 levels, then
just "•" for the remaining levels. And this is just for new bullets,
existing numberings do what's stored in the document.

Change-Id: I4a2a458d3fafe96f702d8c1b2f1cf6cd78b8014d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185208
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>