tdf#130857 qt weld: Take content margins into account
In QtInstanceDrawingArea::handleResizeEvent, take
the QLabel's content margins into account when
setting the new output size for the VirtualDevice
when the QLabel was resized:
The size that the VirtualDevice has available
is not the full label size, but the content margins
need to be subtracted.

Not doing so was seen causing a resize loop for the
ColorPickerDialog (cui/source/dialogs/colorpicker.cxx,
UI file: cui/uiconfig/ui/colorpickerdialog.ui) in
a WIP branch declaring support for that dialog using
native Qt widgets with SAL_VCL_QT_USE_WELDED_WIDGETS=1.

One way to trigger the dialog:

* "Tools" -> "Options"
* go to "Appearance" tab
* in the "Customizations" section, click on the
  color picker next to the "Color" label
* select "Custom Color"

There are more aspects to address before the dialog
can actually be declared as supported.

Change-Id: Iaf183926abd8269297b8ae2d865568cc57fbf2cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185237
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>