tdf#166319 sw interdependent redlines: handle accept of delete under format
The bugdoc has <del>AA<format>BB</format>CC</del> in it, clicking accept
in the middle of AA results in a leftover BB in the document.

This is because the outer redline type of BB is format, so the other
type is not accepted together with the delete accept. But this is not
correct: if the delete is to be accepted, then it's just a small detail
that there was a format on top of it: all this delete is meant to be
accepted, like Word does.

Fix the problem similar to how commit
90077e9b4372e1e3445cc6fb2158a0f2bbb8b1b8 (tdf#166319 sw interdependent
redlines: handle reject of insert under format, 2025-05-09) handled the
reject of inserts: if there is a format, then see if it has an
underlying delete, and if so, still combine the to-be-accepted changes.

Note that rejecting the format + deleting the range means we get
correctly working undo/redo for this use-case.

Change-Id: I8f05d8f94574d5e89432bf9c2df3a737fa8d6d13
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185244
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>