tdf#142011 - Text disappears if "Adjust to contour" selected.
Steps to reproduce - from the bug report:
1. Draw a connector between two objects.
2. Select the connector so that the control points are visible.
3. Click on Insert Text Box to enter text mode.
4. Type the connector text.
5. Right-click on the text and select Text Attributes from the context menu to open the Text dialog.
6. Select "Adjust to contour" and all other options are disabled.
7. Click OK and the Text dialog closes. The text disappears.

And the suggested solution is to hide the mentioned option. Because it is not applicable for certain objects.

While i debug, i saw that also "line" objects are problematic with "Adjust to contour" option. So i also hid the option for line objects.

Signed-off-by: Gökay Şatır <gokaysatir@gmail.com>
Change-Id: I81974eaae66eccc33da66c623256fbcb977ffb42
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185256
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit ad6eaf4559ad4e788dd07ae497a195840ae300ee)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185352
Reviewed-by: Gökay ŞATIR <gokaysatir@collabora.com>
Tested-by: Jenkins