ReqIF: set all necessary borders in merged cells
Commit 63c91b9cb3f73b66a915875721b0efd65b8aebac (sw HTML import: apply
default table autoformat on cells in reqif mode, 2018-08-08) made ReqIF
import apply default table autoformat to tables, because ReqIF has no
cell format info.

Commit 46fcf87725818666fee34226afdac4a64edd94eb (tdf#115576 sw: fix
borders for single row/col autotable styles, 2019-11-19) introduced
special handling for inserting tables with single row and/or column,
because in that case, the applied formatting was for the first (e.g.,
top-left) cell, which lacked bottom/right borders.

In case of ReqIF tables with merged cells, which span till the right
table edge, applying autoformat to such a cell has the same problem:
the cell position is not rightmost, but it needs a right border as in
the rightmost cell.

This change extends the code from the fix of tdf#115576, to handle not
only single-row/single-column cases, but also spanned cells, and fixes
the ReqIF merged cells case.

Change-Id: I50da381d0a74e75e808036bcf335d0857994bf40
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185259
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins