sd: Use concrete rtl::Reference<PresenterPaneBorderPainter>
After

    Change-Id: Iacdd51e0d2d2bde454d66a97d4c468628395a1b5
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue May 13 23:24:27 2025 +0200

        sd: Implement PresenterPaneBase::initialize without XInitialization

, it is now clear that this is always a
PresenterPaneBorderPainter, so use that concrete
type instead of the abstract XPaneBorderPainter
interface.

This also makes clear that the attempt to query
for the XWindow2 interface in
PresenterPaneBase::IsVisible would never succeed,
so (temporarily) simply return false there.
This suggests that PresenterPane::UpdateBoundingBox
which calls PresenterPaneBase::IsVisible might not
be as useful as it looked like at first glance as
it seems to always set `maBoundingBox` to an empty
rectangle...

This will be adjusted in an upcoming commit, but
for now, leave the behavior as it used to be.

Change-Id: I933c63303d1f055b4c0eb826f6726df3fe2a7033
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185288
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins